+++ src/scene/disk.h
@@ -6,16 +6,16 @@
 class Disk : public Geometry
 {
 public:
-	Disk(float radius = 1, const Vec3 &position = Vec3(), const Vec3 &normal = Vec3(0, 1, 0));
+	Disk(double radius = 1, const Vec3 &position = Vec3(), const Vec3 &normal = Vec3(0, 1, 0));
 
-	virtual bool			hit(const Ray &ray, float &t) const;
+	virtual bool			hit(const Ray &ray, double &t) const;
 	virtual void			normalAt(const Vec3 &p, Vec3 &n) const	{ n = this->n; }
 
 private:
 	Vec3	n;
-	float	d;
-	float	r;
-	float	rSquared;
+	double	d;
+	double	r;
+	double	rSquared;
 };
 
 #endif
