+++ src/scene/plane.h
@@ -6,16 +6,16 @@
 class Plane : public Geometry
 {
 public:
-	Plane(const Vec3 &normal = Vec3(0, 1, 0), float offset = 0, bool _oneSided = true)
+	Plane(const Vec3 &normal = Vec3(0, 1, 0), double offset = 0, bool _oneSided = true)
 		: n(normal), d(offset), oneSided(_oneSided)
 	{}
 
-	virtual bool			hit(const Ray &ray, float &t) const;
+	virtual bool			hit(const Ray &ray, double &t) const;
 	virtual void			normalAt(const Vec3 &p, Vec3 &n) const	{ n = this->n; }
 
 protected:
 	Vec3		n;
-	float		d;
+	double		d;
 	bool		oneSided;
 };
