+++ src/scene/sphere.h
@@ -6,14 +6,14 @@
 class Sphere : public Geometry
 {
 public:
-	Sphere(float radius = 1) : r(radius)	{}
-	Sphere(float radius, Vec3 p) : r(radius), Geometry(p) {}
+	Sphere(double radius = 1) : r(radius)	{}
+	Sphere(double radius, Vec3 p) : r(radius), Geometry(p) {}
 
-	virtual bool			hit(const Ray &ray, float &t) const;
+	virtual bool			hit(const Ray &ray, double &t) const;
 	virtual void			normalAt(const Vec3 &p, Vec3 &n) const;
 
 private:
-	float	r;
+	double	r;
 };
 
 #endif
